♻️: refactor AOP 객체 common 모듈로 분리 #9
